fixing transfering history, refactoring
