+++ app/src/main/java/com/romandevyatov/bestfinance/ui/adapters/history/ViewPagerAdapter.kt
@@ -4,15 +4,13 @@ import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.romandevyatov.bestfinance.ui.fragments.history.ExpenseHistoryFragment
 import com.romandevyatov.bestfinance.ui.fragments.history.IncomeHistoryFragment
-import com.romandevyatov.bestfinance.ui.fragments.history.TransferFragment
 import com.romandevyatov.bestfinance.ui.fragments.history.TransferHistoryFragment
-
-
-private const val NUM_TABS = 3
 
 
 class ViewPagerAdapter(fragment: Fragment) :
     FragmentStateAdapter(fragment) {
+
+    private val NUM_TABS = 3
 
     override fun getItemCount(): Int {
         return NUM_TABS
