+++ app/src/main/res/layouts/adapters/layout/transfer_history_card.xml
@@ -20,11 +20,10 @@
             android:layout_alignParentEnd="true"
             android:layout_margin="4dp"
             android:layout_marginStart="4dp"
+            android:text="Transfer Name"
             android:textSize="18sp"
             android:textStyle="bold"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@+id/balanceTextView"
-            app:layout_constraintHorizontal_bias="0.068"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintVertical_bias="1.0" />
@@ -33,11 +32,19 @@
             android:id="@+id/balanceTextView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginEnd="32dp"
+            android:text="Balance"
             android:textSize="16dp"
-            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
+
+        <TextView
+            android:id="@+id/dateTransferTextView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="dateTransferTextView"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/balanceTextView" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
