+++ app/src/main/res/layouts/fragments/layout/fragment_transfer.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_marginStart="10dp"
     android:layout_marginEnd="10dp"
-    tools:context=".ui.fragments.history.TransferFragment">
+    tools:context=".ui.fragments.AddNewTransferFragment">
 
 
     <Spinner
